+++ frontend/apps/user/store/friendSearchForm/friendSearchInputSlice.ts
@@ -1,4 +1,4 @@
-import { FriendSearch } from '@U/components/FriendFrameHeader/FriendSearchInput/FriendSearchInput';
+import { FriendSearch } from '@U/components/FriendFrame/FriendFrameHeader/FriendSearchInput/FriendSearchInput';
 import { UseForm } from '@amadda/react-util-hooks';
 import { ChangeEvent } from 'react';
 import type { StateCreator } from 'zustand';
